pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-blueprint-maker"
version = "1.0.0"
description = "Application to extract individual sprites from sprite sheets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Antigravity" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyQt6",
    "opencv-python",
    "numpy",
    "Pillow",
    "PyOpenGL",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyinstaller",
]

[project.urls]
Homepage = "https://github.com/yuri-schmaltz/my-blueprint-maker"
Repository = "https://github.com/yuri-schmaltz/my-blueprint-maker"

[project.scripts]
my-blueprint-maker = "main:main"

[tool.setuptools]
py-modules = ["main", "main_window", "sprite_extractor", "preview_3d", "extrator_sprites_gimp"]
